Persona = Role
Persona has been cast down into just being roles. Hazzah!
