+++ src/PersonaTester.java
@@ -1,9 +1,9 @@
-import characters.Persona;
+import characters.Role;
 
 public class PersonaTester {
     public static void main(String[] args)
     {
-        Persona newb = new Persona();
+        Role newb = new Role();
         System.out.println(newb);
     }
 }
